- [breaking changes] migrate the functions `shuffleArray` and `shuffledDeque` from `/src/array2d.ts` to `/src/array1d.ts`, and add doc tests to them. - in `/src/array1d.ts`, add a new function `rotateArray`.
